requirements.txt
fastapi[all]
google-auth-oauthlib
google-auth
google-api-python-client
httpx
uvicorn
python-multipart